feat: addressed pr comment
